requirements.txt
streamlit
tensorflow
pandas
numpy
opencv-python
Pillow
scikit-learn